Booking Email Notification
Email notification can only be sent once so the message must be the same
message. I only added a CC so can send to car owner too. I don't believe
we need to send to admin, as the admin himself is sending automated
email messages to owner and renters so you could just get all the emails
from sent items if you are the admin